Remove launder sites the ROM does not need (3/5)
Chunk 3 of 5 of the delaunder sweep, split so each part lands under the
PR validator's 200-file cap and every file gets a real CI byte gate.

176 files. Each site was removed only when the ROM still agreed with the
result, one site at a time, by tools/delaunder.py.

+++ src/func_ov002_020bbb14.cpp
@@ -38,7 +38,7 @@ void func_ov002_020bbb14(char* self)
     unsigned id;
 
     {
-        s16* pa = (s16*)((long long)(int)(self + 0x8c));
+        s16* pa = (s16*)(self + 0x8c);
         *pa = *pa + 0x2000;
     }
     _ZN5Actor9UpdatePosEP12CylinderClsn(self, 0);
