+++ src/func_ov002_020c3a48.c
@@ -20,7 +20,7 @@ int func_ov002_020c3a48(char *self)
         if (_ZNK9Animation12WillHitFrameEi(anim, 0x5a)) {
             _ZN8Particle20RunningSlidingDustAtE5Fix12IiES1_S1_(*(int *)(self + 0x5c), *(int *)(self + 0x60), *(int *)(self + 0x64));
             _ZN5Sound13PlayCharVoiceEjjRK7Vector3(*(unsigned char *)(self + 0x6d9), 0, self + 0x74);
-            *(unsigned char *)(((long long)(int)(self + 0x6e5))) += 1;
+            *(unsigned char *)(self + 0x6e5) += 1;
         }
     } else {
         int buf[6];
@@ -32,7 +32,7 @@ int func_ov002_020c3a48(char *self)
         buf[5] = 0;
         func_ov002_020c37a4(self, (char *)(buf + 3));
         id = _ZNK6Player14GetBodyModelIDEjb(self, *(int *)(self + 8) & 0xff, 0);
-        anim = (char *)(((long long)(int)(*(char **)(self + (id << 2) + 0xdc) + 0x50)));
+        anim = (char *)((long long)(int)(*(char **)(self + (id << 2) + 0xdc) + 0x50));
         r5 = (unsigned int)(*(int *)(anim + 8) << 4) >> 0x10;
         if (r5 == 0x6d || r5 == 0x8d) {
             _ZN8Particle20RunningSlidingDustAtE5Fix12IiES1_S1_(*(int *)(self + 0x5c), *(int *)(self + 0x60), *(int *)(self + 0x64));
@@ -45,7 +45,7 @@ int func_ov002_020c3a48(char *self)
             _ZN5Sound13PlayCharVoiceEjjRK7Vector3(*(unsigned char *)(self + 0x6d9), 4, self + 0x74);
             break;
         }
-        *(int *)(((long long)(int)(self + 0x688))) -= 0x21000;
+        *(int *)(self + 0x688) -= 0x21000;
         if (_ZN6Player12FinishedAnimEv(self)) {
             func_ov002_020c3cf0(self);
             return 1;
