+++ src/func_ov002_020c8f80.cpp
@@ -21,9 +21,9 @@ extern "C" int func_ov002_020c8f80(Player *thiz)
 {
     char *p = (char *)thiz;
 
-    *(int *)(((long long)(int)(p + 0x80))) -= 0x80;
-    *(int *)(((long long)(int)(p + 0x84))) -= 0x80;
-    *(int *)(((long long)(int)(p + 0x88))) -= 0x80;
+    *(int *)(p + 0x80) -= 0x80;
+    *(int *)(p + 0x84) -= 0x80;
+    *(int *)(p + 0x88) -= 0x80;
 
     unsigned int id = thiz->GetBodyModelID(*(int *)(p + 8) & 0xff, false);
     Animation *anim = (Animation *)(*(char **)(p + id * 4 + 0xdc) + 0x50);
