+++ src/func_ov002_020f1fcc.c
@@ -10,13 +10,13 @@ void func_ov002_020f1fcc(unsigned char *self)
     int i, v;
     if (*(unsigned short *)(s + 0x2e) == 0)
         return;
-    (*(unsigned short *)((int)(((long long)(int)(s + 0x2c)))))++;
+    (*(unsigned short *)((int)(s + 0x2c)))++;
     s = *(unsigned char **)(self + 0xd4);
     if (*(unsigned short *)(s + 0x2c) != 0x18)
         return;
     *(unsigned short *)(s + 0x2c) = 0;
     s = *(unsigned char **)(self + 0xd4);
-    (*(unsigned short *)((int)(((long long)(int)(s + 0x2e)))))--;
+    (*(unsigned short *)((int)(s + 0x2e)))--;
     s = *(unsigned char **)(self + 0xd4);
     v = *(unsigned short *)(s + 0x2e);
     if (v != 0) {
@@ -30,7 +30,7 @@ void func_ov002_020f1fcc(unsigned char *self)
     _ZN3G2x13SetBlendAlphaEPVttttt(
         (volatile void *)0x4001050, 0, 0x28, 0xc, 4);
     s = *(unsigned char **)(self + 0xd4);
-    (*(unsigned char *)((int)(((long long)(int)(s + 0x34)))))++;
+    (*(unsigned char *)((int)(s + 0x34)))++;
     for (i = 0; i < 3; i++) {
         s = *(unsigned char **)(self + 0xd4);
         ((unsigned short *)(s + 0x24))[i] = (i + 1) << 4;
