+++ src/func_ov004_020b2adc.c
@@ -18,7 +18,7 @@ void *func_ov004_020b2adc(char *self)
     *(void **)self = &data_0208e4b8;
     *(void **)self = &_ZTV5Scene;
     {
-        unsigned char *p13 = (unsigned char *)(((long long)(int)(self + 0x13)));
+        unsigned char *p13 = (unsigned char *)(self + 0x13);
         *p13 |= 1;
         *p13 |= 4;
     }
