+++ src/func_ov006_020c8084.cpp
@@ -18,7 +18,7 @@ extern int data_ov006_0213b090[2];
 
 extern "C" void func_ov006_020c8084(char *c)
 {
-    int *p = (int*)(((long long)(int)(c + 0x3c)));
+    int *p = (int*)(c + 0x3c);
     int *g = data_ov006_0213b088;
     if (p[0] == g[0] && (p[1] == g[1] || *(int*)(c + 0x3c) == 0)) {
         *(int*)(c + 0x24) = 0;
