+++ src/func_ov006_020d89c4.c
@@ -12,7 +12,7 @@ void func_ov006_020d89c4(char* self)
     func_ov006_020d836c(self);
 
     if (*(u16*)(self + 0x62e8) != 0) {
-        (*(u16*)(((long long)(int)(self + 0x62e8))))--;
+        (*(u16*)(self + 0x62e8))--;
         if (*(s16*)(self + 0x62e8) < 0)
             *(u16*)(self + 0x62e8) = 0;
         return;
@@ -20,7 +20,7 @@ void func_ov006_020d89c4(char* self)
 
     for (j = 0; j < 0x70; j++) {
         u8 *row = (u8*)self + j*0x40;
-        u8 *q = (u8*)(((long long)(int)(row + 0x4698)));
+        u8 *q = (u8*)(row + 0x4698);
         if (*q == 2) {
             if (((u8 (*)[0x40])(self + 0x4000))[j][0x697] == 6
              && ((u8 (*)[0x40])(self + 0x4000))[j][0x69b] == 4) {
